feat!: Change interaction structure and handling of component responses
