+++ src/interactions/commands/player/stop.ts
@@ -2,26 +2,25 @@ import config from 'config';
 import { NodeResolvable, useQueue } from 'discord-player';
 import { EmbedBuilder, GuildMember, SlashCommandBuilder } from 'discord.js';
 
-import loggerModule from '../../services/logger';
-import { CommandParams } from '../../types/commandTypes';
-import { EmbedOptions } from '../../types/configTypes';
-import { queueDoesNotExist, queueIsEmpty } from '../../utils/validation/queueValidator';
-import { notInSameVoiceChannel, notInVoiceChannel } from '../../utils/validation/voiceChannelValidator';
+import loggerModule from '../../../services/logger';
+import { CommandParams } from '../../../types/commandTypes';
+import { EmbedOptions } from '../../../types/configTypes';
+import { notInSameVoiceChannel, notInVoiceChannel } from '../../../utils/validation/voiceChannelValidator';
 
 const embedOptions: EmbedOptions = config.get('embedOptions');
 module.exports = {
     isNew: false,
     isBeta: false,
     data: new SlashCommandBuilder()
-        .setName('shuffle')
-        .setDescription('Shuffle tracks in the queue randomly.')
+        .setName('stop')
+        .setDescription('Stop playing audio and clear the track queue.')
         .setDMPermission(false)
         .setNSFW(false),
     execute: async ({ interaction, executionId }: CommandParams) => {
         const logger = loggerModule.child({
-            source: 'shuffle.js',
+            source: 'stop.js',
             module: 'slashCommand',
-            name: '/shuffle',
+            name: '/stop',
             executionId: executionId,
             shardId: interaction.guild?.shardId,
             guildId: interaction.guild?.id
@@ -33,20 +32,31 @@ module.exports = {
 
         const queue: NodeResolvable = useQueue(interaction.guild!.id)!;
 
-        if (await queueDoesNotExist({ interaction, queue, executionId })) {
-            return;
+        if (!queue) {
+            logger.debug('There is no queue.');
+
+            logger.debug('Responding with warning embed.');
+            return await interaction.editReply({
+                embeds: [
+                    new EmbedBuilder()
+                        .setDescription(
+                            `**${embedOptions.icons.warning} Oops!**\n_Hmm.._ It seems I am not in a voice channel!`
+                        )
+                        .setColor(embedOptions.colors.warning)
+                ]
+            });
         }
 
         if (await notInSameVoiceChannel({ interaction, queue, executionId })) {
             return;
         }
 
-        if (await queueIsEmpty({ interaction, queue, executionId })) {
-            return;
+        if (!queue.deleted) {
+            queue.setRepeatMode(0);
+            queue.clear();
+            queue.node.stop();
+            logger.debug('Cleared and stopped the queue.');
         }
-
-        queue.tracks.shuffle();
-        logger.debug('Shuffled queue tracks.');
 
         let authorName: string;
 
@@ -65,7 +75,7 @@ module.exports = {
                         iconURL: interaction.user.avatarURL() || ''
                     })
                     .setDescription(
-                        `**${embedOptions.icons.shuffled} Shuffled queue tracks**\nThe **${queue.tracks.data.length}** tracks in the queue has been shuffled.\n\nView the new queue order with **\`/queue\`**.`
+                        `**${embedOptions.icons.success} Stopped playing**\nStopped playing audio and cleared the track queue.\n\nTo play more music, use the **\`/play\`** command!`
                     )
                     .setColor(embedOptions.colors.success)
             ]
