+++ src/types/interactionTypes.ts
@@ -1,0 +1,7 @@
+import { MessageComponentInteraction } from 'discord.js';
+
+export interface ComponentInteractionParams {
+    interaction: MessageComponentInteraction;
+    trackId?: string;
+    executionId: string;
+}
